gengine modified to follow natural JavaScript style
Got read of new and other pseudo-object-oriented artifacts.
The intention is now to modify the whole code not to try to
imitate the object-oriented desing, but to use the most
natural and efficent JavaScript style.

Not bad, for many programmers it takes one year of JavaScript
programming to realize how weird object-oriented design is
when simulated in JavaScript.